docker_images/deepdr/requirements.txt
# DeepDR Requirements
# Deep learning framework
tensorboard

# Required for Registry (cross-model encoder support)
rdkit>=2021.09.1
tqdm>=4.62.0

# Data processing
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
PyYAML>=5.4
joblib>=1.0.0

# Model specific
lifelines>=0.27.0
prettytable>=2.0.0

# Scientific computing
scipy>=1.7.0
scikit-learn>=0.24.0

# PyTorch Geometric - installed separately in Dockerfile (requires PyTorch first)
# torch-geometric, torch-scatter, torch-sparse

# LangChain for run_model_execution.py
langchain>=0.0.1
langchain-core>=0.0.1
python-dotenv>=0.19.0
langchain-mcp-adapters>=0.0.1

# Optional for development
jupyter
matplotlib
seaborn
